+++ src/main/java/com/hyerijang/dailypay/auth/service/AuthenticationService.java
@@ -4,12 +4,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.hyerijang.dailypay.auth.dto.AuthenticationRequest;
 import com.hyerijang.dailypay.auth.dto.AuthenticationResponse;
 import com.hyerijang.dailypay.auth.dto.RegisterRequest;
+import com.hyerijang.dailypay.auth.token.Token;
+import com.hyerijang.dailypay.auth.token.TokenRepository;
+import com.hyerijang.dailypay.auth.token.TokenType;
 import com.hyerijang.dailypay.common.exception.ApiException;
 import com.hyerijang.dailypay.common.exception.response.ExceptionEnum;
-import com.hyerijang.dailypay.config.JwtService;
-import com.hyerijang.dailypay.token.Token;
-import com.hyerijang.dailypay.token.TokenRepository;
-import com.hyerijang.dailypay.token.TokenType;
 import com.hyerijang.dailypay.user.domain.User;
 import com.hyerijang.dailypay.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
