+++ src/main/java/com/hyerijang/dailypay/auth/token/Token.java
@@ -1,4 +1,4 @@
-package com.hyerijang.dailypay.token;
+package com.hyerijang.dailypay.auth.token;
 
 import com.hyerijang.dailypay.user.domain.User;
 import jakarta.persistence.CascadeType;
