+++ src/main/java/com/hyerijang/dailypay/auth/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.hyerijang.dailypay.auth.token;
+
+public enum TokenType {
+    BEARER
+}
