--- src/main/java/com/hyerijang/dailypay/token/TokenType.java
@@ -1,5 +1,0 @@
-package com.hyerijang.dailypay.token;
-
-public enum TokenType {
-    BEARER
-}
